# src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Notes Management API",
    "version": "1.0.0",
    "description": "Secure multi-user notes management service backend API."
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Authenticate user and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid email or password" }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "Get all notes for authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of user notes" }
        }
      },
      "post": {
        "summary": "Create a new note",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" }
                },
                "required": ["title", "content"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Note created successfully" }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "summary": "Get a specific note by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Note details" },
          "404": { "description": "Note not found" }
        }
      },
      "put": {
        "summary": "Update an existing note",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Note updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No content" }
        }
      }
    },
    "/notes/{id}/share": {
      "post": {
        "summary": "Share a note with another user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "share_with_email": { "type": "string" }
                },
                "required": ["share_with_email"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Note shared successfully" }
        }
      }
    },
    "/about": {
      "get": {
        "summary": "About information for the assignment",
        "responses": {
          "200": { "description": "About info payload" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}